Use compiler-rt's unorddf2 for x86_64 and arm64
Bug: http://b/28149048

Clang generates call to __unorddf2 builtin when compiling s_fmal.c in
bionic.  Generate this builtin from compiler-rt for arm64 and x86_64
(where the corresponding libgcc does not have this builtin.

This change is in a separate CL than the other Android.mk updates so
this can also be cherry-picked to aosp/master.

Test: Tested build, boot and common usage for Arm, Arm64, x86, x86_64,
Mips images in AOSP and internal branch.

Change-Id: Iea28c53ba2241cf7e671eed4eea704808d68ea95
Signed-off-by: Eliminater74 <eliminater74@gmail.com>